feat: Find more options
Directs people to files in cmfive where more options can be found for the part of their module they are creating. That is cmfive settings that can be used and cmfive fixed columns that all tables will automatically have, and further precoded html elements that can be added in the template.

refs: 10186