Making code sample fit the format of other samples
All other code samples within the tutorial have the `<?php` opening tag on it's own line followed by a line of white space. Auto-format when saving the document in the codespace also puts the `<?php` on it's own line. To my knowledge, no other code sample in this tutorial uses the PHP close tag, so I removed that. I also changed the codeblock to use PHP syntax highlighting as opposed to HTML highlighting.